feat: add advanced cross-border trade feature.
this commit allows proper calculation when cross-border trade is
enabled.

In short:

* price for end-customers should stay the same no matter the tax
* for business customers, VAT should be removed.
* VAT amount should be calculated based on destination country default
  tax rate and NOT default store tax rate like in magento default.

Current implementation is hacky and brings duplicaiton, this is mostly
caused by the location of the code we need to override here. Target code
is located in:

    Magento\Tax\Model\Calculation\AbstractCalculator

that class is inherited by whole bunch of classes:

    Magento\Tax\Model\Calculation\AbstractAggregateCalculator
    Magento\Tax\Model\Calculation\RowBaseCalculator
    Magento\Tax\Model\Calculation\TotalBaseCalculator
    Magento\Tax\Model\Calculation\UnitBaseCalculator
    Vertex\Tax\Model\Calculation\VertexCalculator

most interesting for us here is RowBaseCalculator and
TotalBaseCalculator. We also override UnitBaseCalculator, though I'm not
sure if that is needed.
I do not override VertexCalculator for 2 reasons:

* I didn't test this
* somebody who uses Vertex platform will not bother implementing proper
  tax calculation in magento itself, as Vertex will calculate tax.

Possible future improvement here is to rewrite implementation so that
required change is not duplicated and injected either as Plugin or at
least as Trait. Leaving current implementation is also considered a good
option though, cause it heavily relies on initialized dependencies of
`AbstractCalculator` some of those dependencies are tricky to get
initialized in the plugin, e.g. `storeId` or `getAddressRateRequest()`.

More generally I think this solution can be implemented best by patching
magento core, author is actually surprised why magento does not support
such calculation mechanics out of the box.